Major Bugfix: Buffer overrun in renewAddress
It was right there the whole time! This painful bug took forever to
find, and it sooo shouldn't have...  sizeof(contactNode) is 8, not 2...

This bug has existed since
https://github.com/TMRh20/RF24Mesh/commit/590d4d64e71c1331f5ccb8b867ccb313267c18df
on Dec 1, 2015 and caused all kinds of mayhem, since it was overwriting
a memory space that changes with changes to the code, so symptoms seemed
vary significantly.

On RPi, segfaults and odd behavior were the main issues. This will
affect all platforms, but only on RPi was it really noticeable so far.